fix tests (and move jemalloc to own file)
